requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
garminconnect==0.2.25
garth==0.5.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
lxml==5.3.1
mcp==1.3.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.0.1
readchar==4.2.1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
withings-sync==4.2.5
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.126.0
